backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
pydantic==2.5.2
python-multipart==0.0.6
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
python-dotenv==1.0.0
chromadb==0.4.22 